requirements.txt
ipython==8.12.3
jedi==0.19.1
jieba
joblib==1.4.2
json5==0.9.25
opencv_python==4.9.0.80
oss2==2.18.6
parso==0.8.4
PyYAML==6.0.1
streamlit==1.36.0
sympy==1.13.1
tiktoken==0.7.0
torch==2.4.0
transformers==4.42.4
zhipuai==1.0.7
dashscope>=1.20.1
faiss-cpu
grpcio
json5
jupyter>=1.0.0
langchain
langchain-community
langchain-experimental
llama-index
llama-index-readers-json
llama-index-retrievers-bm25==0.1.5
llama-index-embeddings-huggingface
modelscope[framework]>=1.16.0
openai
opencv-python
openpyxl
pdfminer.six
Pillow
pydantic>=2.3.0
pytest
pytest-mock
python-dotenv
requests>=2.32.3
seaborn
sentencepiece
tiktoken
unstructured
uvicorn
fastapi
sqlmodel
modelscope_agent